fixed pc compiler warning

+++ yalnix_framework/src/kernel_start.c
@@ -8,12 +8,13 @@
 #include "kernel_memory.h"
 #include "load_program.h"
 #include "pcb.h"
+#include "pipe.h"
 #include "scheduler.h"
+#include "stdint.h"
+#include "synchronization.h"
 #include "tty.h"
 #include "user_memory.h"
-#include "synchronization.h"
-#include "pipe.h"
-#include <assert.h>
+#include <stdint.h>
 
 /*
  * Kernel startup and bootstrap logic.
@@ -30,10 +31,8 @@ static void do_idle(void) {
   }
 }
 
-
-
 static void set_up_uc(UserContext *uc, void (*idle_func)(void), void *sp) {
-  uc->pc = (void *)idle_func;
+  uc->pc = (void *)(uintptr_t)idle_func;
   uc->sp = sp;
 }
 
@@ -68,7 +67,7 @@ void KernelStart(char *cmd_args[], unsigned int pmem_size, UserContext *uctxt) {
 
   pcb_t *curr_proc = get_running_proc();
 
-  char* init_name = cmd_args[0];
+  char *init_name = cmd_args[0];
 
   if (curr_proc == init_pcb) {
     if (LoadProgram(init_name, cmd_args, curr_proc) != SUCCESS)
